scripts/requirements.txt
matplotlib
numpy
pandas
torch 
torchvision 
torchaudio
pytorch-lightning
pyro-ppl 
ConfigSpace
smac